+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.4.0 - 2022-05-09
 
-- [](): feat: adds `withPageAuth`
+- [](): feat: adds `withPageAuth` [docs]() and deprecates `withAuthRequired`.
+- [](): feat: adds `withApiAuth` [docs]() and deprecates `withAuthRequired`.
 
 - [BREAKING CHANGE][](): fix: remove `@supabase/ui` as dependency. If you are using the Auth component, please make sure to install `@supabase/ui` separately in your project and import `Auth` from there. Fixes #39; Fixes #64; Fixes #69;